+++ src/anygeometry/__init__.py
@@ -321,4 +321,4 @@ __all__ = [
     "JunctionMemberUse",
 ]
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"